refactor: Member 수정
