+++ src/main/java/member/domain/Member.java
@@ -1,4 +1,4 @@
-package member;
+package member.domain;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
