+++ src/test/java/member/MemberTest.java
@@ -1,10 +1,10 @@
 package member;
 
-import static org.junit.jupiter.api.Assertions.*;
-
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import member.domain.Member;
 
 class MemberTest {
